LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-url-block.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="../../resources/frame-ancestors-test.js"></script>
</head>
<body>
    <script>
        description("A 'frame-ancestors' CSP directive with a URL which doesn't match this origin should be blocked.");

        crossOriginFrameShouldBeBlocked("http://example.com/");
    </script>
</body>
</html>
