LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-url-block-expected.txt

CONSOLE MESSAGE: Refused to display 'http://localhost:8080/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=http://example.com/' in a frame because an ancestor violates the following Content Security Policy directive: "frame-ancestors http://example.com/".
A 'frame-ancestors' CSP directive with a URL which doesn't match this origin should be blocked.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


IFrame load event fired: the IFrame is cross-origin (or was blocked).
PASS The IFrame should have been blocked (or cross-origin). It was.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
